requirements.txt
Django==1.8.2
django-filter==1.0.0
djangorestframework==3.5.3
psycopg2==2.6.2